Changes to be committed: 	new file:   Block.cpp 	new file:   Block.h 	Added header for the block class, and an implementation file 	Using a vector instead of dynamic array in Block class
	modified:   README.md
	new file:   main
	modified:   main.cpp
